back/app/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn==0.30.3
pydantic-settings==2.4.0
python-multipart==0.0.20
SQLAlchemy==2.0.36
redis==5.0.4
orjson==3.10.7
pyJWT==2.10.0
httpx==0.28.1
backoff==2.2.1
werkzeug==3.1.3
pytz==2025.2

# Database
alembic==1.16.5
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Elasticsearch
elasticsearch==8.19.1
elasticsearch-dsl==8.18.0
aiohttp==3.12.15

# Admin panel
sqladmin==0.21.0
itsdangerous==2.2.0

# Additional utilities
python-dateutil==2.9.0